Mutate subset of data

Like mutate-table.express but for a single call and some .SDcols.

mutate_sd(.data, .how = identity, ..., .SDcols,
  .parse = getOption("table.express.parse", FALSE))

Arguments

.data
An instance of ExprBuilder.

.how
The transmuting call.

...
More arguments for .how if it is a function or a function-call.

.SDcols
See data.table::data.table and the details here.

.parse
Logical. Whether to apply rlang::parse_expr() to obtain the expressions.

Details

This function works similar to transmute_sd() but keeps all columns and can modify by reference, like mutate-table.express.

Additionally, .SDcols supports tidyselect::select_helpers, with the caveat that the expression is evaluated eagerly, i.e. with the currently captured data.table. Consider using chain() to explicitly capture intermediate results as actual data.tables.

Examples

data("mtcars")

data.table::as.data.table(mtcars) %>%
    start_expr %>%
    mutate_sd(.COL * 2, .SDcols = c("mpg", "cyl")) %>%
    end_expr

#>      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#>  1: 42.0  12 160.0 110 3.90 2.620 16.46  0  1    4    4
#>  2: 42.0  12 160.0 110 3.90 2.875 17.02  0  1    4    4
#>  3: 45.6   8 108.0  93 3.85 2.320 18.61  1  1    4    1
#>  4: 42.8  12 258.0 110 3.08 3.215 19.44  1  0    3    1
#>  5: 37.4  16 360.0 175 3.15 3.440 17.02  0  0    3    2
#>  6: 36.2  12 225.0 105 2.76 3.460 20.22  1  0    3    1
#>  7: 28.6  16 360.0 245 3.21 3.570 15.84  0  0    3    4
#>  8: 48.8   8 146.7  62 3.69 3.190 20.00  1  0    4    2
#>  9: 45.6   8 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 10: 38.4  12 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 11: 35.6  12 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 12: 32.8  16 275.8 180 3.07 4.070 17.40  0  0    3    3
#> 13: 34.6  16 275.8 180 3.07 3.730 17.60  0  0    3    3
#> 14: 30.4  16 275.8 180 3.07 3.780 18.00  0  0    3    3
#> 15: 20.8  16 472.0 205 2.93 5.250 17.98  0  0    3    4
#> 16: 20.8  16 460.0 215 3.00 5.424 17.82  0  0    3    4
#> 17: 29.4  16 440.0 230 3.23 5.345 17.42  0  0    3    4
#> 18: 64.8   8  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 19: 60.8   8  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 20: 67.8   8  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 21: 43.0   8 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 22: 31.0  16 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 23: 30.4  16 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 24: 26.6  16 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 25: 38.4  16 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 26: 54.6   8  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 27: 52.0   8 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 28: 60.8   8  95.1 113 3.77 1.513 16.90  1  1    5    2
#> 29: 31.6  16 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 30: 39.4  12 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 31: 30.0  16 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 32: 42.8   8 121.0 109 4.11 2.780 18.60  1  1    4    2
#>      mpg cyl  disp  hp drat    wt  qsec vs am gear carb