Order by clause

Clause for ordering rows.

order_by(.data, ...)

# S3 method for ExprBuilder
order_by(
  .data,
  ...,
  .collapse,
  .parse = getOption("table.express.parse", FALSE),
  .chain = getOption("table.express.chain", TRUE)
)

# S3 method for data.table
order_by(.data, ...)

Arguments

.data

The input data.

...

Arguments for the specific methods.

.collapse

Ignored. See details.

.parse

Logical. Whether to apply rlang::parse_expr() to obtain the expressions.

.chain

Logical. Should a new frame be automatically chained to the expression if the clause being set already exists?

Details

The ExprBuilder method dispatches to where-table.express, but doesn't forward the .collapse argument.

To see more examples, check the vignette, or the table.express-package entry.

Examples

data("mtcars")

data.table::as.data.table(mtcars) %>%
    order_by(-cyl, gear)
#>      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#>  1: 18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#>  2: 14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#>  3: 16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#>  4: 17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#>  5: 15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#>  6: 10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#>  7: 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#>  8: 14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#>  9: 15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> 10: 15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> 11: 13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> 12: 19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> 13: 15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> 14: 15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> 15: 21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> 16: 18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> 17: 21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> 18: 21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> 19: 19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> 20: 17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> 21: 19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> 22: 21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> 23: 22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> 24: 24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> 25: 22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> 26: 32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> 27: 30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> 28: 33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> 29: 27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> 30: 21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 31: 26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> 32: 30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#>      mpg cyl  disp  hp drat    wt  qsec vs am gear carb