Frame expression builder

Build an expression that will be used inside a data.table::data.table's frame. This shouldn't be used directly.

Methods

initialize(DT, dt_pronouns = list(), .verbose)

Constructor that receives a data.table::data.table in DT. The dt_pronouns parameter is used internally when chaining for joins.

set_j(value, chain_if_needed)

Set the j clause expression(s), starting a new frame if the current one already has said expression set.

set_i(value, chain_if_needed)

Like set_j but for the i clause.

set_by(value, chain_if_needed)

Set the by clause expression.

chain(type = "frame", dt)

By default, start a new expression with the current one as its parent. If type = "pronoun", dt is used to start a new expression that joins the current one.

eval(parent_env, by_ref, ...)

Evaluate the final expression with parent_env as the enclosing environment. If by_ref = FALSE, data.table::copy() is called before. The ellipsis' contents are assigned to the expression's evaluation environment.

tidy_select(select_expr)

Evaluate a tidyselect call using the currently captured table.

print(...)

Prints the built expr.

Active bindings

appends

Extra expressions that go at the end.

expr

The final expression that can be evaluated with base::eval() or rlang::eval_bare().

Methods

Public methods

ExprBuilder$new()

ExprBuilder$set_i()

ExprBuilder$set_j()

ExprBuilder$set_by()

ExprBuilder$chain()

ExprBuilder$chain_if_set()

ExprBuilder$seek_and_nestroy()

ExprBuilder$eval()

ExprBuilder$tidy_select()

ExprBuilder$print()

ExprBuilder$clone()

Method new()

Usage

ExprBuilder$new(
  DT,
  dt_pronouns = list(),
  nested = list(),
  verbose = getOption("table.express.verbose", FALSE)
)

Method set_i()

Usage

ExprBuilder$set_i(value, chain_if_needed)

Method set_j()

Usage

ExprBuilder$set_j(value, chain_if_needed)

Method set_by()

Usage

ExprBuilder$set_by(value, chain_if_needed)

Method chain()

Usage

ExprBuilder$chain(type = "frame", next_dt, parent_env)

Method chain_if_set()

Usage

ExprBuilder$chain_if_set(...)

Method seek_and_nestroy()

Usage

ExprBuilder$seek_and_nestroy(.exprs)

Method eval()

Usage

ExprBuilder$eval(parent_env, by_ref, ...)

Method tidy_select()

Usage

ExprBuilder$tidy_select(select_expr)

Method print()

Usage

ExprBuilder$print(...)

Method clone()

The objects of this class are cloneable with this method.

Usage

ExprBuilder$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.