Filter subset of data

Helper to filter rows with the same condition applied to a subset of the data.

filter_sd(.data, .SDcols, .how = Negate(is.na), ..., .collapse = `&`,
  .parse = getOption("table.express.parse", FALSE),
  .chain = getOption("table.express.chain", TRUE))

Arguments

.data
An instance of ExprBuilder.

.SDcols
See data.table::data.table and the details here.

.how
The filtering function or predicate.

...
Possibly more arguments for .how.

.collapse
See where-table.express.

.parse
Logical. Whether to apply rlang::parse_expr() to obtain the expressions.

.chain
Logical. Should a new frame be automatically chained to the expression if the clause being set already exists?

Details

This function adds/chains an i expression that will be evaluated by data.table::data.table, and it supports the .COL pronoun. The .how condition is applied to all .SDcols.

Additionally, .SDcols supports:

tidyselect::select_helpers

A predicate using the .COL pronoun that should return a single logical when .COL is replaced by a column of the data.

The caveat is that the expression is evaluated eagerly, i.e. with the currently captured data.table. Consider using chain() to explicitly capture intermediate results as actual data.tables.

To see more examples, check the vignette, or the table.express-package entry.

Examples

data("mtcars")

data.table::as.data.table(mtcars) %>%
    start_expr %>%
    filter_sd(c("vs", "am"), .COL == 1)

#> .DT_[vs == 1 & am == 1]