Joining verbs

Two-table joins. Check the "Joining verbs" vignette for more information.

# S3 method for ExprBuilder
anti_join(x, y, ...)

# S3 method for ExprBuilder
full_join(x, y, ..., sort = TRUE, allow = TRUE,
  .parent_env)

# S3 method for ExprBuilder
inner_join(x, y, ...)

# S3 method for ExprBuilder
left_join(x, y, ..., nomatch, mult, roll, rollends,
  .parent_env)

mutate_join(x, y, ...)

# S3 method for ExprBuilder
mutate_join(x, y, ..., .SDcols, mult, roll, rollends,
  allow = FALSE, .by_each = NULL, .parent_env)

# S3 method for ExprBuilder
right_join(x, y, ..., nomatch, mult, roll, rollends)

# S3 method for ExprBuilder
semi_join(x, y, ..., .parent_env)

Arguments

x
An ExprBuilder instance.

y
A data.table::data.table.

...
Expressions for the on part of the join.

sort
Passed to data.table::merge.

allow
Passed as data.table's allow.cartesian.

.parent_env
See end_expr().

nomatch
See data.table::data.table.

mult
See data.table::data.table.

roll
See data.table::data.table.

rollends
See data.table::data.table.

.SDcols
For mutate_join. See the details below.

.by_each
For mutate_join. See the details below.

Details

The full_join method is really a wrapper for data.table::merge that specifies all = TRUE. The expression in x gets evaluated, merged with y, and the result is captured in a new ExprBuilder. Useful in case you want to keep building expressions after the merge.

The semi_join method starts a new expression immediately.

Mutating join

The ExprBuilder method for mutate_join implements the idiom described in this link. The columns specified in .SDcols are those that will be added to x from y. The specification can be done by:

Using tidyselect::select_helpers.

Passing a character vector. If the character is named, the names are taken as the new column names for the values added to x.

A list, using base::list() or .(), containing:

Column names, either as characters or symbols.

Named calls expressing how the column should be summarized/modified before adding it to x.

The last case mentioned above is useful when the join returns many rows from y for each row in x, so they can be summarized while joining. The value of by in the join depends on what is passed to .by_each:

If NULL (the default), by is set to .EACHI if a call is detected in any of the expressions from the list in .SDcols

If TRUE, by is always set to .EACHI

If FALSE, by is never set to .EACHI

See also

Examples

lhs <- data.table::data.table(x = rep(c("b", "a", "c"), each = 3),
                              y = c(1, 3, 6),
                              v = 1:9)

rhs <- data.table::data.table(x = c("c", "b"),
                              v = 8:7,
                              foo = c(4, 2))

rhs %>%
    start_expr %>%
    anti_join(lhs, x, v) %>%
    end_expr

#>    x v foo
#> 1: b 7   2

lhs %>%
    start_expr %>%
    inner_join(rhs, x) %>%
    end_expr

#>    x y v i.v foo
#> 1: c 1 7   8   4
#> 2: c 3 8   8   4
#> 3: c 6 9   8   4
#> 4: b 1 1   7   2
#> 5: b 3 2   7   2
#> 6: b 6 3   7   2

# creates new data.table
lhs %>%
    start_expr %>%
    left_join(rhs, x) %>%
    end_expr

#>    x  v foo y i.v
#> 1: b  7   2 1   1
#> 2: b  7   2 3   2
#> 3: b  7   2 6   3
#> 4: a NA  NA 1   4
#> 5: a NA  NA 3   5
#> 6: a NA  NA 6   6
#> 7: c  8   4 1   7
#> 8: c  8   4 3   8
#> 9: c  8   4 6   9

# would modify lhs by reference
lhs %>%
    start_expr %>%
    mutate_join(rhs, x, .SDcols = c("foo", rhs.v = "v"))

#> .DT_[, `:=`(c("foo", "rhs.v"), rhs[.SD, list(foo = x.foo, rhs.v = x.v), 
#>     on = list(x)])]

# would modify rhs by reference, summarizing 'y' before adding it.
rhs %>%
    start_expr %>%
    mutate_join(lhs, x, .SDcols = .(y = mean(y)))

#> .DT_[, `:=`("y", lhs[.SD, list(y = mean(y)), by = .EACHI, on = list(x)][, 
#>     list(y)])]

# creates new data.table
lhs %>%
    start_expr %>%
    right_join(rhs, x) %>%
    end_expr

#>    x y v i.v foo
#> 1: c 1 7   8   4
#> 2: c 3 8   8   4
#> 3: c 6 9   8   4
#> 4: b 1 1   7   2
#> 5: b 3 2   7   2
#> 6: b 6 3   7   2

# keep only columns from lhs
lhs %>%
    start_expr %>%
    semi_join(rhs, x) %>%
    end_expr

#>    x y v
#> 1: b 1 1
#> 2: b 3 2
#> 3: b 6 3
#> 4: c 1 7
#> 5: c 3 8
#> 6: c 6 9