Transmute subset of data

Like transmute-table.express but for a single call and maybe specifying .SDcols.

transmute_sd(.data, .how = identity, ..., .SDcols = names(.SD),
  .parse = getOption("table.express.parse", FALSE), .chain = TRUE)

Arguments

.data
An instance of ExprBuilder.

.how
The transmuting call.

...
More arguments for .how if it is a function or a function-call.

.SDcols
See data.table::data.table and the details here.

.parse
Logical. Whether to apply rlang::parse_expr() to obtain the expressions.

.chain
Logical. Should a new frame be automatically chained to the expression if the clause being set already exists?

Details

Like transmute-table.express, this function never modifies the input by reference. However, this function does operate on ExprBuilder instances, and adds/chains a select expression that will be evaluated by data.table::data.table, possibly specifying the helper function .transmute_matching, which is assigned to the final expression's evaluation environment when calling end_expr() (i.e., ExprBuilder's eval method).

Said function includes two pronouns that can be used by .how and .SDcols:

.COL: the actual values of the column.

.COLNAME: the name of the column currently being evaluated.

Unlike a call like DT[, (vars) := expr], .SDcols can be created dynamically with an expression that evaluates to something that would be used in place of vars. See the examples.

Examples

data("mtcars")

data.table::as.data.table(mtcars) %>%
    start_expr %>%
    transmute_sd(.COL * 2, .SDcols = grepl("^d", .COLNAME)) %>%
    end_expr

#>      disp drat
#>  1: 320.0 7.80
#>  2: 320.0 7.80
#>  3: 216.0 7.70
#>  4: 516.0 6.16
#>  5: 720.0 6.30
#>  6: 450.0 5.52
#>  7: 720.0 6.42
#>  8: 293.4 7.38
#>  9: 281.6 7.84
#> 10: 335.2 7.84
#> 11: 335.2 7.84
#> 12: 551.6 6.14
#> 13: 551.6 6.14
#> 14: 551.6 6.14
#> 15: 944.0 5.86
#> 16: 920.0 6.00
#> 17: 880.0 6.46
#> 18: 157.4 8.16
#> 19: 151.4 9.86
#> 20: 142.2 8.44
#> 21: 240.2 7.40
#> 22: 636.0 5.52
#> 23: 608.0 6.30
#> 24: 700.0 7.46
#> 25: 800.0 6.16
#> 26: 158.0 8.16
#> 27: 240.6 8.86
#> 28: 190.2 7.54
#> 29: 702.0 8.44
#> 30: 290.0 7.24
#> 31: 602.0 7.08
#> 32: 242.0 8.22
#>      disp drat